refactor: restructure Extension Host backend architecture
- Split YoloDataReader god class into ConfigLoader/ImageFileScanner/LabelCodec
- Remove UiService and ImageService, replace with function modules webviewHtml.ts and imageLoader.ts
- Simplify ErrorHandler (remove dead code like error history and auto-recovery)
- Merge WebviewMessageHandler duplicate handlers (next/previous, openImage/openTxt)
- Simplify CacheManager to simple LRU cache
