+++ src/LabelingPanel.ts
@@ -1,12 +1,11 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
-import { YoloDataReader } from './YoloDataReader';
+import { YoloDataReader } from './yolo/YoloDataReader';
 import { ErrorHandler, ErrorType } from './ErrorHandler';
-import { CacheManager } from './model/CacheManager';
 import { WebviewMessage } from './model/types';
-import { ImageService } from './services/ImageService';
-import { UiService } from './services/UiService';
 import { WebviewMessageHandler } from './services/WebviewMessageHandler';
+import { getErrorHtml, generateWebviewHtml } from './utils/webviewHtml';
+import { loadImageAsDataUrl } from './utils/imageLoader';
 
 export class LabelingPanel {
     // Store multiple panel instances in a Map with the YAML path as key
@@ -17,10 +16,7 @@ export class LabelingPanel {
     private readonly _yamlUri: vscode.Uri;
     private _disposables: vscode.Disposable[] = [];
     private _yoloReader?: YoloDataReader;
-    private _imageCache: CacheManager<string>;
     private _hasError: boolean = false;
-    private _imageService: ImageService;
-    private _uiService: UiService;
     private _messageHandler!: WebviewMessageHandler;
 
     public static createOrShow(extensionUri: vscode.Uri, yamlUri: vscode.Uri) {
@@ -69,20 +65,7 @@ export class LabelingPanel {
         this._panel = panel;
         this._extensionUri = extensionUri;
         this._yamlUri = yamlUri;
-        
-        // 创建缓存管理器，使用更具描述性的配置参数
-        const CACHE_CONFIG = {
-            maxItems: 20,             // 最大缓存项数
-            maxSize: 20 * 1024,       // 最大缓存大小（KB）
-            ttl: 10 * 60 * 1000       // 缓存生存时间（毫秒）
-        };
-        
-        this._imageCache = new CacheManager<string>(CACHE_CONFIG);
-        
-        // 初始化服务类
-        this._imageService = new ImageService(this._imageCache);
-        this._uiService = new UiService(this._extensionUri);
-        
+
         try {
             const workspaceFolder = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
             this._yoloReader = new YoloDataReader(yamlUri.fsPath, workspaceFolder);
@@ -93,9 +76,8 @@ export class LabelingPanel {
         }
 
         this._messageHandler = new WebviewMessageHandler(
-            this._panel.webview, 
-            this._yoloReader,
-            this._imageService
+            this._panel.webview,
+            this._yoloReader
         );
         
         this._update();
@@ -110,7 +92,7 @@ export class LabelingPanel {
     private _validateYoloReader(): void {
         if (!this._yoloReader?.getClassNames().length || !this._yoloReader?.getCurrentImage()) {
             this._hasError = true;
-            this._panel.webview.html = this._uiService.getErrorHtml('No images or classes found in dataset');
+            this._panel.webview.html = getErrorHtml('No images or classes found in dataset');
         }
     }
 
@@ -131,7 +113,7 @@ export class LabelingPanel {
             }
         );
         
-        this._panel.webview.html = this._uiService.getErrorHtml(`Failed to load dataset configuration: ${error.message}`);
+        this._panel.webview.html = getErrorHtml(`Failed to load dataset configuration: ${error.message}`);
     }
 
     private async _update() {
@@ -141,7 +123,7 @@ export class LabelingPanel {
 
     private async _getHtmlForWebview(webview: vscode.Webview) {
         if (!this._yoloReader) {
-            return this._uiService.getErrorHtml('YoloDataReader is not initialized');
+            return getErrorHtml('YoloDataReader is not initialized');
         }
         
         const classNames = this._yoloReader.getClassNames();
@@ -153,18 +135,21 @@ export class LabelingPanel {
             initialImageData = currentImage ? await this._loadImage(currentImage) : null;
         } catch (error: any) {
             this._hasError = true;
-            return this._uiService.getErrorHtml(`Failed to load image: ${error.message}`);
+            return getErrorHtml(`Failed to load image: ${error.message}`);
         }
 
-        // 获取图像计数信息
+        // 获取图像计数信息，与 WebviewMessageHandler.getImageInfoText 保持格式一致
         const totalImages = this._yoloReader.getTotalImages();
         const currentIndex = this._yoloReader.getCurrentImageIndex();
-        const imageInfoText = `Image: ${currentIndex + 1} of ${totalImages}`;
+        const filename = this._yoloReader.getCurrentImage()
+            ? path.basename(this._yoloReader.getCurrentImage()!)
+            : '';
+        const imageInfoText = `图片 ${currentIndex + 1}/${totalImages} - ${filename}`;
 
         // 获取关键点配置
         const kptShape = this._yoloReader.getKptShape();
 
-        return this._uiService.generateWebviewHtml({
+        return generateWebviewHtml(this._extensionUri, {
             classNames,
             initialImageData,
             initialLabels: labels,
@@ -203,7 +188,7 @@ export class LabelingPanel {
     }
 
     private async _loadImage(imagePath: string): Promise<string> {
-        return this._imageService.loadImage(imagePath);
+        return loadImageAsDataUrl(imagePath);
     }
 
     public dispose() {
